Ref cleaner return and file renames
